kafka: move rewriter to its own file
