+++ contrib/kafka/filters/network/source/broker/filter.cc
@@ -69,37 +69,6 @@ absl::flat_hash_map<int32_t, MonotonicTime>& KafkaMetricsFacadeImpl::getRequestA
   return request_arrivals_;
 }
 
-ResponseRewriter::ResponseRewriter(const BrokerFilterConfig& config) : config_{config} {};
-
-void ResponseRewriter::onMessage(AbstractResponseSharedPtr response) {
-  if (config_.force_response_rewrite_) {
-    responses_to_rewrite_.push_back(response);
-  }
-}
-
-void ResponseRewriter::onFailedParse(ResponseMetadataSharedPtr) { /* Nothing to do. */
-}
-
-void ResponseRewriter::rewrite(Buffer::Instance& buffer) {
-  if (!config_.force_response_rewrite_) {
-    // Data received from upstream is going to be sent downstream without any changes.
-    return;
-  }
-
-  // At this stage we have access to responses, and can do something to them.
-  buffer.drain(buffer.length());
-  ResponseEncoder encoder{buffer};
-  ENVOY_LOG(trace, "emitting {} stored responses", responses_to_rewrite_.size());
-  for (auto response : responses_to_rewrite_) {
-    encoder.encode(*response);
-  }
-  responses_to_rewrite_.erase(responses_to_rewrite_.begin(), responses_to_rewrite_.end());
-}
-
-size_t ResponseRewriter::getStoredResponseCountForTest() const {
-  return responses_to_rewrite_.size();
-}
-
 KafkaBrokerFilter::KafkaBrokerFilter(Stats::Scope& scope, TimeSource& time_source,
                                      const BrokerFilterConfig& filter_config)
     : KafkaBrokerFilter{filter_config, std::make_shared<KafkaMetricsFacadeImpl>(
